Rename and Consolidate EventLoops
Motivation:

Pull Request #2681 modified several EventLoops and EventLoopGroups
to no longer manage its threads directly but instead
relinquish control of threading and scheduling to an Executor.

This commit adjusts some of the class names accordingly.

Modifications:

At first:
* AbstractEventLoop was merged with EmbeddedEventLoop.
* AbstractEventExecutor was merged with EmbeddedEventLoop, ImmediateEventExecutor, GlobalEventExecutor and SingleThreadEventExecutor.

Then, AbstractEvent(Loop|Executor) and AbstractEvent(Loop|Executor)Group were deleted.

Finally, the following classes were renamed:
* SingleThreadEvent(Loop|Executor) -> AbstractEvent(Loop|Executor)
* MultithreadEvent(Loop|Executor)Group -> AbstractEvent(Loop|Executor)Group

The system property io.netty.eventLoopThreads was renamed to io.netty.eventLoopParallelism.

Result:

More sensible class names and cleaner inheritance tree.
