+++ common/src/main/java/io/netty/util/concurrent/AbstractEventExecutor.java
@@ -15,30 +15,917 @@
  */
 package io.netty.util.concurrent;
 
+import io.netty.util.internal.CallableEventExecutorAdapter;
+import io.netty.util.internal.PlatformDependent;
+import io.netty.util.internal.RunnableEventExecutorAdapter;
+import io.netty.util.internal.logging.InternalLogger;
+import io.netty.util.internal.logging.InternalLoggerFactory;
+
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Iterator;
+import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.PriorityQueue;
+import java.util.Queue;
 import java.util.Set;
 import java.util.concurrent.AbstractExecutorService;
+import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.Callable;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.RunnableFuture;
+import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
+
+import static io.netty.util.concurrent.AbstractEventExecutorGroup.*;
 
 /**
- * Abstract base class for {@link EventExecutor} implementations.
+ * Abstract base class for {@link EventExecutor}'s that execute all its submitted tasks in a single thread.
+ *
  */
 public abstract class AbstractEventExecutor extends AbstractExecutorService implements EventExecutor {
 
-    static final long DEFAULT_SHUTDOWN_QUIET_PERIOD = 2;
-    static final long DEFAULT_SHUTDOWN_TIMEOUT = 15;
+    private static final InternalLogger logger =
+            InternalLoggerFactory.getInstance(AbstractEventExecutor.class);
+
+    private static final int ST_NOT_STARTED = 1;
+    private static final int ST_STARTED = 2;
+    private static final int ST_SHUTTING_DOWN = 3;
+    private static final int ST_SHUTDOWN = 4;
+    private static final int ST_TERMINATED = 5;
+
+    private static final Runnable WAKEUP_TASK = new Runnable() {
+        @Override
+        public void run() {
+            // Do nothing.
+        }
+    };
+
+    private static final AtomicIntegerFieldUpdater<AbstractEventExecutor> STATE_UPDATER;
+
+    private static final long threadOffset;
+
+    static {
+        AtomicIntegerFieldUpdater<AbstractEventExecutor> updater =
+                PlatformDependent.newAtomicIntegerFieldUpdater(AbstractEventExecutor.class, "state");
+        if (updater == null) {
+            updater = AtomicIntegerFieldUpdater.newUpdater(AbstractEventExecutor.class, "state");
+        }
+        STATE_UPDATER = updater;
+
+        try {
+            threadOffset =
+                    PlatformDependent.objectFieldOffset(AbstractEventExecutor.class.getDeclaredField("thread"));
+        } catch (NoSuchFieldException e) {
+            throw new RuntimeException();
+        }
+    }
 
     private final EventExecutorGroup parent;
 
-    protected AbstractEventExecutor() {
-        this(null);
+    private final Queue<Runnable> taskQueue;
+    final Queue<ScheduledFutureTask<?>> delayedTaskQueue = new PriorityQueue<ScheduledFutureTask<?>>();
+
+    @SuppressWarnings({ "FieldMayBeFinal", "unused" })
+    private volatile Thread thread;
+    private final Executor executor;
+    private final Semaphore threadLock = new Semaphore(0);
+    private final Set<Runnable> shutdownHooks = new LinkedHashSet<Runnable>();
+    private final boolean addTaskWakesUp;
+
+    private long lastExecutionTime;
+
+    @SuppressWarnings({ "FieldMayBeFinal", "unused" })
+    private volatile int state = ST_NOT_STARTED;
+
+    private volatile long gracefulShutdownQuietPeriod;
+    private volatile long gracefulShutdownTimeout;
+    private long gracefulShutdownStartTime;
+
+    private final Promise<?> terminationFuture = new DefaultPromise<Void>(GlobalEventExecutor.INSTANCE);
+
+    private boolean firstRun = true;
+
+    private final Runnable asRunnable = new Runnable() {
+        @Override
+        public void run() {
+            updateThread(Thread.currentThread());
+
+            // lastExecutionTime must be set on the first run
+            // in order for shutdown to work correctly for the
+            // rare case that the eventloop did not execute
+            // a single task during its lifetime.
+            if (firstRun) {
+                firstRun = false;
+                updateLastExecutionTime();
+            }
+
+            try {
+                AbstractEventExecutor.this.run();
+            } catch (Throwable t) {
+                logger.warn("Unexpected exception from an event executor: ", t);
+                cleanupAndTerminate(false);
+            }
+        }
+    };
+
+    /**
+     * @param parent            the {@link EventExecutorGroup} which is the parent of this instance and belongs to it.
+     * @param executor          the {@link Executor} which will be used for executing.
+     * @param addTaskWakesUp    {@code true} if and only if invocation of {@link #addTask(Runnable)} will wake up
+     *                          the executor thread.
+     */
+    protected AbstractEventExecutor(EventExecutorGroup parent, Executor executor, boolean addTaskWakesUp) {
+        if (executor == null) {
+            throw new NullPointerException("executor");
+        }
+
+        this.parent = parent;
+        this.addTaskWakesUp = addTaskWakesUp;
+        this.executor = executor;
+        taskQueue = newTaskQueue();
     }
 
-    protected AbstractEventExecutor(EventExecutorGroup parent) {
-        this.parent = parent;
+    /**
+     * Create a new {@link Queue} which will holds the tasks to execute. This default implementation will return a
+     * {@link LinkedBlockingQueue} but if your sub-class of {@link AbstractEventExecutor} will not do any blocking
+     * calls on the this {@link Queue} it may make sense to {@code @Override} this and return some more performant
+     * implementation that does not support blocking operations at all.
+     */
+    protected Queue<Runnable> newTaskQueue() {
+        return new LinkedBlockingQueue<Runnable>();
+    }
+
+    /**
+     * @see {@link Queue#poll()}
+     */
+    protected Runnable pollTask() {
+        assert inEventLoop();
+        for (;;) {
+            Runnable task = taskQueue.poll();
+            if (task == WAKEUP_TASK) {
+                continue;
+            }
+            return task;
+        }
+    }
+
+    /**
+     * Take the next {@link Runnable} from the task queue and so will block if no task is currently present.
+     * <p>
+     * Be aware that this method will throw an {@link UnsupportedOperationException} if the task queue, which was
+     * created via {@link #newTaskQueue()}, does not implement {@link BlockingQueue}.
+     * </p>
+     *
+     * @return {@code null} if the executor thread has been interrupted or waken up.
+     */
+    protected Runnable takeTask() {
+        assert inEventLoop();
+        if (!(taskQueue instanceof BlockingQueue)) {
+            throw new UnsupportedOperationException();
+        }
+
+        BlockingQueue<Runnable> taskQueue = (BlockingQueue<Runnable>) this.taskQueue;
+        for (;;) {
+            ScheduledFutureTask<?> delayedTask = delayedTaskQueue.peek();
+            if (delayedTask == null) {
+                Runnable task = null;
+                try {
+                    task = taskQueue.take();
+                    if (task == WAKEUP_TASK) {
+                        task = null;
+                    }
+                } catch (InterruptedException e) {
+                    // Ignore
+                }
+                return task;
+            } else {
+                long delayNanos = delayedTask.delayNanos();
+                Runnable task = null;
+                if (delayNanos > 0) {
+                    try {
+                        task = taskQueue.poll(delayNanos, TimeUnit.NANOSECONDS);
+                    } catch (InterruptedException e) {
+                        // Waken up.
+                        return null;
+                    }
+                }
+                if (task == null) {
+                    // We need to fetch the delayed tasks now as otherwise there may be a chance that
+                    // delayed tasks are never executed if there is always one task in the taskQueue.
+                    // This is for example true for the read task of OIO Transport
+                    // See https://github.com/netty/netty/issues/1614
+                    fetchFromDelayedQueue();
+                    task = taskQueue.poll();
+                }
+
+                if (task != null) {
+                    return task;
+                }
+            }
+        }
+    }
+
+    private void fetchFromDelayedQueue() {
+        long nanoTime = 0L;
+        for (;;) {
+            ScheduledFutureTask<?> delayedTask = delayedTaskQueue.peek();
+            if (delayedTask == null) {
+                break;
+            }
+
+            if (nanoTime == 0L) {
+                nanoTime = ScheduledFutureTask.nanoTime();
+            }
+
+            if (delayedTask.deadlineNanos() <= nanoTime) {
+                delayedTaskQueue.remove();
+                taskQueue.add(delayedTask);
+            } else {
+                break;
+            }
+        }
+    }
+
+    /**
+     * @see {@link Queue#peek()}
+     */
+    protected Runnable peekTask() {
+        assert inEventLoop();
+        return taskQueue.peek();
+    }
+
+    /**
+     * @see {@link Queue#isEmpty()}
+     */
+    protected boolean hasTasks() {
+        assert inEventLoop();
+        return !taskQueue.isEmpty();
+    }
+
+    /**
+     * Returns {@code true} if a scheduled task is ready for processing by {@link #runAllTasks()} or
+     * {@link #runAllTasks(long)}.
+     */
+    protected boolean hasScheduledTasks() {
+        assert inEventLoop();
+        ScheduledFutureTask<?> delayedTask = delayedTaskQueue.peek();
+        return delayedTask != null && delayedTask.deadlineNanos() <= ScheduledFutureTask.nanoTime();
+    }
+
+    /**
+     * Return the number of tasks that are pending for processing.
+     *
+     * <strong>Be aware that this operation may be expensive as it depends on the internal implementation of the
+     * AbstractEventExecutor. So use it was care!</strong>
+     */
+    public final int pendingTasks() {
+        return taskQueue.size();
+    }
+
+    /**
+     * Add a task to the task queue, or throws a {@link RejectedExecutionException} if this instance was shutdown
+     * before.
+     */
+    protected void addTask(Runnable task) {
+        if (task == null) {
+            throw new NullPointerException("task");
+        }
+        if (isShutdown()) {
+            reject();
+        }
+        taskQueue.add(task);
+    }
+
+    /**
+     * @see {@link Queue#remove(Object)}
+     */
+    protected boolean removeTask(Runnable task) {
+        if (task == null) {
+            throw new NullPointerException("task");
+        }
+        return taskQueue.remove(task);
+    }
+
+    /**
+     * Poll all tasks from the task queue and run them via {@link Runnable#run()} method.
+     *
+     * @return {@code true} if and only if at least one task was run
+     */
+    protected boolean runAllTasks() {
+        fetchFromDelayedQueue();
+        Runnable task = pollTask();
+        if (task == null) {
+            return false;
+        }
+
+        for (;;) {
+            try {
+                task.run();
+            } catch (Throwable t) {
+                logger.warn("A task raised an exception.", t);
+            }
+
+            task = pollTask();
+            if (task == null) {
+                lastExecutionTime = ScheduledFutureTask.nanoTime();
+                return true;
+            }
+        }
+    }
+
+    /**
+     * Poll all tasks from the task queue and run them via {@link Runnable#run()} method.  This method stops running
+     * the tasks in the task queue and returns if it ran longer than {@code timeoutNanos}.
+     */
+    protected boolean runAllTasks(long timeoutNanos) {
+        fetchFromDelayedQueue();
+        Runnable task = pollTask();
+        if (task == null) {
+            return false;
+        }
+
+        final long deadline = ScheduledFutureTask.nanoTime() + timeoutNanos;
+        long runTasks = 0;
+        long lastExecutionTime;
+        for (;;) {
+            try {
+                task.run();
+            } catch (Throwable t) {
+                logger.warn("A task raised an exception.", t);
+            }
+
+            runTasks ++;
+
+            // Check timeout every 64 tasks because nanoTime() is relatively expensive.
+            // XXX: Hard-coded value - will make it configurable if it is really a problem.
+            if ((runTasks & 0x3F) == 0) {
+                lastExecutionTime = ScheduledFutureTask.nanoTime();
+                if (lastExecutionTime >= deadline) {
+                    break;
+                }
+            }
+
+            task = pollTask();
+            if (task == null) {
+                lastExecutionTime = ScheduledFutureTask.nanoTime();
+                break;
+            }
+        }
+
+        this.lastExecutionTime = lastExecutionTime;
+        return true;
+    }
+
+    /**
+     * Returns the amount of time left until the scheduled task with the closest dead line is executed.
+     */
+    protected long delayNanos(long currentTimeNanos) {
+        ScheduledFutureTask<?> delayedTask = delayedTaskQueue.peek();
+        if (delayedTask == null) {
+            return SCHEDULE_PURGE_INTERVAL;
+        }
+
+        return delayedTask.delayNanos(currentTimeNanos);
+    }
+
+    /**
+     * Updates the internal timestamp that tells when a submitted task was executed most recently.
+     * {@link #runAllTasks()} and {@link #runAllTasks(long)} updates this timestamp automatically, and thus there's
+     * usually no need to call this method.  However, if you take the tasks manually using {@link #takeTask()} or
+     * {@link #pollTask()}, you have to call this method at the end of task execution loop for accurate quiet period
+     * checks.
+     */
+    protected void updateLastExecutionTime() {
+        lastExecutionTime = ScheduledFutureTask.nanoTime();
+    }
+
+    /**
+     *
+     */
+    protected abstract void run();
+
+    /**
+     * Do nothing, sub-classes may override
+     */
+    protected void cleanup() {
+        // NOOP
+    }
+
+    protected void wakeup(boolean inEventLoop) {
+        if (!inEventLoop || STATE_UPDATER.get(this) == ST_SHUTTING_DOWN) {
+            taskQueue.add(WAKEUP_TASK);
+        }
+    }
+
+    @Override
+    public boolean inEventLoop(Thread thread) {
+        return thread == this.thread;
+    }
+
+    /**
+     * Add a {@link Runnable} which will be executed on shutdown of this instance
+     */
+    public void addShutdownHook(final Runnable task) {
+        if (inEventLoop()) {
+            shutdownHooks.add(task);
+        } else {
+            execute(new Runnable() {
+                @Override
+                public void run() {
+                    shutdownHooks.add(task);
+                }
+            });
+        }
+    }
+
+    /**
+     * Remove a previous added {@link Runnable} as a shutdown hook
+     */
+    public void removeShutdownHook(final Runnable task) {
+        if (inEventLoop()) {
+            shutdownHooks.remove(task);
+        } else {
+            execute(new Runnable() {
+                @Override
+                public void run() {
+                    shutdownHooks.remove(task);
+                }
+            });
+        }
+    }
+
+    private boolean runShutdownHooks() {
+        boolean ran = false;
+        // Note shutdown hooks can add / remove shutdown hooks.
+        while (!shutdownHooks.isEmpty()) {
+            List<Runnable> copy = new ArrayList<Runnable>(shutdownHooks);
+            shutdownHooks.clear();
+            for (Runnable task: copy) {
+                try {
+                    task.run();
+                } catch (Throwable t) {
+                    logger.warn("Shutdown hook raised an exception.", t);
+                } finally {
+                    ran = true;
+                }
+            }
+        }
+
+        if (ran) {
+            lastExecutionTime = ScheduledFutureTask.nanoTime();
+        }
+
+        return ran;
+    }
+
+    @Override
+    public Future<?> shutdownGracefully(long quietPeriod, long timeout, TimeUnit unit) {
+        if (quietPeriod < 0) {
+            throw new IllegalArgumentException("quietPeriod: " + quietPeriod + " (expected >= 0)");
+        }
+        if (timeout < quietPeriod) {
+            throw new IllegalArgumentException(
+                    "timeout: " + timeout + " (expected >= quietPeriod (" + quietPeriod + "))");
+        }
+        if (unit == null) {
+            throw new NullPointerException("unit");
+        }
+
+        if (isShuttingDown()) {
+            return terminationFuture();
+        }
+
+        boolean inEventLoop = inEventLoop();
+        boolean wakeup;
+        int oldState;
+        for (;;) {
+            if (isShuttingDown()) {
+                return terminationFuture();
+            }
+            int newState;
+            wakeup = true;
+            oldState = STATE_UPDATER.get(this);
+            if (inEventLoop) {
+                newState = ST_SHUTTING_DOWN;
+            } else {
+                switch (oldState) {
+                    case ST_NOT_STARTED:
+                    case ST_STARTED:
+                        newState = ST_SHUTTING_DOWN;
+                        break;
+                    default:
+                        newState = oldState;
+                        wakeup = false;
+                }
+            }
+            if (STATE_UPDATER.compareAndSet(this, oldState, newState)) {
+                break;
+            }
+        }
+        gracefulShutdownQuietPeriod = unit.toNanos(quietPeriod);
+        gracefulShutdownTimeout = unit.toNanos(timeout);
+
+        if (oldState == ST_NOT_STARTED) {
+            scheduleExecution();
+        }
+
+        if (wakeup) {
+            wakeup(inEventLoop);
+        }
+
+        return terminationFuture();
+    }
+
+    @Override
+    public Future<?> terminationFuture() {
+        return terminationFuture;
+    }
+
+    @Override
+    @Deprecated
+    public void shutdown() {
+        if (isShutdown()) {
+            return;
+        }
+
+        boolean inEventLoop = inEventLoop();
+        boolean wakeup;
+        int oldState;
+        for (;;) {
+            if (isShuttingDown()) {
+                return;
+            }
+            int newState;
+            wakeup = true;
+            oldState = STATE_UPDATER.get(this);
+            if (inEventLoop) {
+                newState = ST_SHUTDOWN;
+            } else {
+                switch (oldState) {
+                    case ST_NOT_STARTED:
+                    case ST_STARTED:
+                    case ST_SHUTTING_DOWN:
+                        newState = ST_SHUTDOWN;
+                        break;
+                    default:
+                        newState = oldState;
+                        wakeup = false;
+                }
+            }
+            if (STATE_UPDATER.compareAndSet(this, oldState, newState)) {
+                break;
+            }
+        }
+
+        if (oldState == ST_NOT_STARTED) {
+            scheduleExecution();
+        }
+
+        if (wakeup) {
+            wakeup(inEventLoop);
+        }
+    }
+
+    @Override
+    public boolean isShuttingDown() {
+        return STATE_UPDATER.get(this) >= ST_SHUTTING_DOWN;
+    }
+
+    @Override
+    public boolean isShutdown() {
+        return STATE_UPDATER.get(this) >= ST_SHUTDOWN;
+    }
+
+    @Override
+    public boolean isTerminated() {
+        return STATE_UPDATER.get(this) == ST_TERMINATED;
+    }
+
+    /**
+     * Confirm that the shutdown if the instance should be done now!
+     */
+    protected boolean confirmShutdown() {
+        if (!isShuttingDown()) {
+            return false;
+        }
+
+        if (!inEventLoop()) {
+            throw new IllegalStateException("must be invoked from an event loop");
+        }
+
+        cancelDelayedTasks();
+
+        if (gracefulShutdownStartTime == 0) {
+            gracefulShutdownStartTime = ScheduledFutureTask.nanoTime();
+        }
+
+        if (runAllTasks() || runShutdownHooks()) {
+            if (isShutdown()) {
+                // Executor shut down - no new tasks anymore.
+                return true;
+            }
+
+            // There were tasks in the queue. Wait a little bit more until no tasks are queued for the quiet period.
+            wakeup(true);
+            return false;
+        }
+
+        final long nanoTime = ScheduledFutureTask.nanoTime();
+
+        if (isShutdown() || nanoTime - gracefulShutdownStartTime > gracefulShutdownTimeout) {
+            return true;
+        }
+
+        if (nanoTime - lastExecutionTime <= gracefulShutdownQuietPeriod) {
+            // Check if any tasks were added to the queue every 100ms.
+            // TODO: Change the behavior of takeTask() so that it returns on timeout.
+            wakeup(true);
+            try {
+                Thread.sleep(100);
+            } catch (InterruptedException e) {
+                // Ignore
+            }
+
+            return false;
+        }
+
+        // No tasks were added for last quiet period - hopefully safe to shut down.
+        // (Hopefully because we really cannot make a guarantee that there will be no execute() calls by a user.)
+        return true;
+    }
+
+    private void cancelDelayedTasks() {
+        if (delayedTaskQueue.isEmpty()) {
+            return;
+        }
+
+        final ScheduledFutureTask<?>[] delayedTasks =
+                delayedTaskQueue.toArray(new ScheduledFutureTask<?>[delayedTaskQueue.size()]);
+
+        for (ScheduledFutureTask<?> task: delayedTasks) {
+            task.cancel(false);
+        }
+
+        delayedTaskQueue.clear();
+    }
+
+    @Override
+    public boolean awaitTermination(long timeout, TimeUnit unit) throws InterruptedException {
+        if (unit == null) {
+            throw new NullPointerException("unit");
+        }
+
+        if (inEventLoop()) {
+            throw new IllegalStateException("cannot await termination of the current thread");
+        }
+
+        if (threadLock.tryAcquire(timeout, unit)) {
+            threadLock.release();
+        }
+
+        return isTerminated();
+    }
+
+    @Override
+    public void execute(Runnable task) {
+        if (task == null) {
+            throw new NullPointerException("task");
+        }
+
+        boolean inEventLoop = inEventLoop();
+        if (inEventLoop) {
+            addTask(task);
+        } else {
+            startExecution();
+            addTask(task);
+            if (isShutdown() && removeTask(task)) {
+                reject();
+            }
+        }
+
+        if (!addTaskWakesUp && wakesUpForTask(task)) {
+            wakeup(inEventLoop);
+        }
+    }
+
+    @SuppressWarnings("unused")
+    protected boolean wakesUpForTask(Runnable task) {
+        return true;
+    }
+
+    protected static void reject() {
+        throw new RejectedExecutionException("event executor terminated");
+    }
+
+    // ScheduledExecutorService implementation
+
+    private static final long SCHEDULE_PURGE_INTERVAL = TimeUnit.SECONDS.toNanos(1);
+
+    @Override
+    public ScheduledFuture<?> schedule(Runnable command, long delay, TimeUnit unit) {
+        if (command == null) {
+            throw new NullPointerException("command");
+        }
+        if (unit == null) {
+            throw new NullPointerException("unit");
+        }
+        if (delay < 0) {
+            throw new IllegalArgumentException(
+                    String.format("delay: %d (expected: >= 0)", delay));
+        }
+        return schedule(new ScheduledFutureTask<Void>(
+                this, delayedTaskQueue, toCallable(command), ScheduledFutureTask.deadlineNanos(unit.toNanos(delay))));
+    }
+
+    @Override
+    public <V> ScheduledFuture<V> schedule(Callable<V> callable, long delay, TimeUnit unit) {
+        if (callable == null) {
+            throw new NullPointerException("callable");
+        }
+        if (unit == null) {
+            throw new NullPointerException("unit");
+        }
+        if (delay < 0) {
+            throw new IllegalArgumentException(
+                    String.format("delay: %d (expected: >= 0)", delay));
+        }
+        return schedule(new ScheduledFutureTask<V>(
+                this, delayedTaskQueue, callable, ScheduledFutureTask.deadlineNanos(unit.toNanos(delay))));
+    }
+
+    @Override
+    public ScheduledFuture<?> scheduleAtFixedRate(Runnable command, long initialDelay, long period, TimeUnit unit) {
+        if (command == null) {
+            throw new NullPointerException("command");
+        }
+        if (unit == null) {
+            throw new NullPointerException("unit");
+        }
+        if (initialDelay < 0) {
+            throw new IllegalArgumentException(
+                    String.format("initialDelay: %d (expected: >= 0)", initialDelay));
+        }
+        if (period <= 0) {
+            throw new IllegalArgumentException(
+                    String.format("period: %d (expected: > 0)", period));
+        }
+
+        return schedule(new ScheduledFutureTask<Void>(
+                this, delayedTaskQueue, toCallable(command),
+                ScheduledFutureTask.deadlineNanos(unit.toNanos(initialDelay)), unit.toNanos(period)));
+    }
+
+    @Override
+    public ScheduledFuture<?> scheduleWithFixedDelay(Runnable command, long initialDelay, long delay, TimeUnit unit) {
+        if (command == null) {
+            throw new NullPointerException("command");
+        }
+        if (unit == null) {
+            throw new NullPointerException("unit");
+        }
+        if (initialDelay < 0) {
+            throw new IllegalArgumentException(
+                    String.format("initialDelay: %d (expected: >= 0)", initialDelay));
+        }
+        if (delay <= 0) {
+            throw new IllegalArgumentException(
+                    String.format("delay: %d (expected: > 0)", delay));
+        }
+
+        return schedule(new ScheduledFutureTask<Void>(
+                this, delayedTaskQueue, toCallable(command),
+                ScheduledFutureTask.deadlineNanos(unit.toNanos(initialDelay)), -unit.toNanos(delay)));
+    }
+
+    private <V> ScheduledFuture<V> schedule(final ScheduledFutureTask<V> task) {
+        if (task == null) {
+            throw new NullPointerException("task");
+        }
+
+        if (inEventLoop()) {
+            delayedTaskQueue.add(task);
+        } else {
+            execute(new Runnable() {
+                @Override
+                public void run() {
+                    delayedTaskQueue.add(task);
+                }
+            });
+        }
+
+        return task;
+    }
+
+    private static Callable<Void> toCallable(final Runnable command) {
+        if (command instanceof RunnableEventExecutorAdapter) {
+            return new RunnableToCallableAdapter((RunnableEventExecutorAdapter) command);
+        } else {
+            return Executors.<Void>callable(command, null);
+        }
+    }
+
+    protected void cleanupAndTerminate(boolean success) {
+        for (;;) {
+            int oldState = STATE_UPDATER.get(AbstractEventExecutor.this);
+            if (oldState >= ST_SHUTTING_DOWN || STATE_UPDATER.compareAndSet(
+                    AbstractEventExecutor.this, oldState, ST_SHUTTING_DOWN)) {
+                break;
+            }
+        }
+
+        // Check if confirmShutdown() was called at the end of the loop.
+        if (success && gracefulShutdownStartTime == 0) {
+            logger.error("Buggy " + EventExecutor.class.getSimpleName() + " implementation; " +
+                    AbstractEventExecutor.class.getSimpleName() + ".confirmShutdown() must be called " +
+                    "before run() implementation terminates.");
+        }
+
+        try {
+            // Run all remaining tasks and shutdown hooks.
+            for (;;) {
+                if (confirmShutdown()) {
+                    break;
+                }
+            }
+        } finally {
+            try {
+                cleanup();
+            } finally {
+                STATE_UPDATER.set(AbstractEventExecutor.this, ST_TERMINATED);
+                threadLock.release();
+                if (!taskQueue.isEmpty()) {
+                    logger.warn(
+                            "An event executor terminated with " +
+                                    "non-empty task queue (" + taskQueue.size() + ')');
+                }
+
+                firstRun = true;
+                terminationFuture.setSuccess(null);
+            }
+        }
+    }
+
+    private void startExecution() {
+        if (STATE_UPDATER.get(this) == ST_NOT_STARTED) {
+            if (STATE_UPDATER.compareAndSet(this, ST_NOT_STARTED, ST_STARTED)) {
+                delayedTaskQueue.add(new ScheduledFutureTask<Void>(
+                        this, delayedTaskQueue, Executors.<Void>callable(new PurgeTask(), null),
+                        ScheduledFutureTask.deadlineNanos(SCHEDULE_PURGE_INTERVAL), -SCHEDULE_PURGE_INTERVAL));
+                scheduleExecution();
+            }
+        }
+    }
+
+    protected final void scheduleExecution() {
+        updateThread(null);
+        executor.execute(asRunnable);
+    }
+
+    private void updateThread(Thread t) {
+        PlatformDependent.putOrderedObject(this, threadOffset, t);
+    }
+
+    private final class PurgeTask implements Runnable {
+        @Override
+        public void run() {
+            Iterator<ScheduledFutureTask<?>> i = delayedTaskQueue.iterator();
+            while (i.hasNext()) {
+                ScheduledFutureTask<?> task = i.next();
+                if (task.isCancelled()) {
+                    i.remove();
+                }
+            }
+        }
+    }
+
+    private static final class RunnableToCallableAdapter implements CallableEventExecutorAdapter<Void> {
+
+        final RunnableEventExecutorAdapter runnable;
+
+        RunnableToCallableAdapter(RunnableEventExecutorAdapter runnable) {
+            this.runnable = runnable;
+        }
+
+        @Override
+        public EventExecutor executor() {
+            return runnable.executor();
+        }
+
+        @Override
+        public Callable unwrap() {
+            return null;
+        }
+
+        @Override
+        public Void call() throws Exception {
+            runnable.run();
+            return null;
+        }
     }
 
     @Override
@@ -71,13 +958,6 @@ public abstract class AbstractEventExecutor extends AbstractExecutorService impl
     public Future<?> shutdownGracefully() {
         return shutdownGracefully(DEFAULT_SHUTDOWN_QUIET_PERIOD, DEFAULT_SHUTDOWN_TIMEOUT, TimeUnit.SECONDS);
     }
-
-    /**
-     * @deprecated {@link #shutdownGracefully(long, long, TimeUnit)} or {@link #shutdownGracefully()} instead.
-     */
-    @Override
-    @Deprecated
-    public abstract void shutdown();
 
     /**
      * @deprecated {@link #shutdownGracefully(long, long, TimeUnit)} or {@link #shutdownGracefully()} instead.
@@ -132,26 +1012,5 @@ public abstract class AbstractEventExecutor extends AbstractExecutorService impl
     @Override
     protected final <T> RunnableFuture<T> newTaskFor(Callable<T> callable) {
         return new PromiseTask<T>(this, callable);
-    }
-
-    @Override
-    public ScheduledFuture<?> schedule(Runnable command, long delay,
-                                       TimeUnit unit) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <V> ScheduledFuture<V> schedule(Callable<V> callable, long delay, TimeUnit unit) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public ScheduledFuture<?> scheduleAtFixedRate(Runnable command, long initialDelay, long period, TimeUnit unit) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public ScheduledFuture<?> scheduleWithFixedDelay(Runnable command, long initialDelay, long delay, TimeUnit unit) {
-        throw new UnsupportedOperationException();
     }
 }
