+++ common/src/main/java/io/netty/util/concurrent/DefaultEventExecutor.java
@@ -18,10 +18,10 @@ package io.netty.util.concurrent;
 import java.util.concurrent.Executor;
 
 /**
- * Default {@link SingleThreadEventExecutor} implementation which just execute all submitted task in a
+ * Default {@link AbstractEventExecutor} implementation which just execute all submitted task in a
  * serial fashion.
  */
-public final class DefaultEventExecutor extends SingleThreadEventExecutor {
+public final class DefaultEventExecutor extends AbstractEventExecutor {
 
     public DefaultEventExecutor() {
         this((EventExecutorGroup) null);
